chore: 메인 파일과 관련 함수 파일의 분리
